pyproject.toml
[tool.poetry]
name = "attack-service"
version = "0.1.0"
description = "Service for scanning linked VMs for weak items"
authors = ["Andrei <sharik.and@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11.2"
fastapi =  {extras = ["all"], version = "^0.95.2"}
python-dotenv = "^1.0.0"
requests = "^2.31.0"
motor = "^3.1.2"


[tool.poetry.group.dev.dependencies]
parameterized = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
